fixed: move deliveryAction back into DeliveryTrait
